Auth for Description II

+++ src/controllers/descriptions.controller.ts
@@ -31,6 +31,7 @@ const descriptionData = (Description: description): queryParam => [
 ];
 
 class DescriptionsController {
+    private type = 1;
     async get(id?: number, selector = 'id'): Promise<dataResponse> {
         const options: options = { where: { selector, value: id } };
         return await db.select('*', id ? options : null);
@@ -39,11 +40,10 @@ class DescriptionsController {
     async post(Description: description, auth?:string): Promise<dataResponse> {
         const {brand, color, dimensions, other, product_id} = Description;
 
-        const user_id = await this.getUserId(product_id);
+        const user_id = await this.getUserId({productId:product_id});
 
         
-        if(auth && await jwtCtrl.checkToken(auth, {id:user_id, type:1})){
-            console.log(await jwtCtrl.checkToken(auth, {id:user_id, type:1}));
+        if(auth && await jwtCtrl.checkToken(auth, {id:user_id, type:this.type})){
             if (brand|| color|| dimensions|| other|| product_id)
                 return await db.insert(descriptionData(Description), {
                     toValidate: { selector: 'product_id' },
@@ -56,26 +56,45 @@ class DescriptionsController {
     async update(id: number, Description: description, auth?:string): Promise<dataResponse> {
         if(Description.id || Description.product_id)  return errors.idCannotChange;
         const {brand, color, dimensions, other} = Description;
-        if (brand|| color|| dimensions|| other)
-            return await db.update(descriptionData(Description), {
-                selector: 'id',
-                value: id,
-            });
-        else return errors.requestEmpty;
+        const user_id = await this.getUserId({descId:id});
+
+        if(auth && jwtCtrl.checkToken(auth, {id:user_id, type:this.type})){
+            if (brand|| color|| dimensions|| other)
+                return await db.update(descriptionData(Description), {
+                    selector: 'id',
+                    value: id,
+                });
+            else return errors.requestEmpty;
+        }
+        return errors.notAuth;
     }
 
     async delete(id: number, auth?:string): Promise<dataResponse> {
-        return await db.delete({ selector: 'id', value: id });
+        if(auth && jwtCtrl.checkToken(auth, {id:await this.getUserId({descId:id})}))
+            return await db.delete({ selector: 'id', value: id });
+        return errors.notAuth;
     }
 
-    private async getUserId(productId:number):Promise<number>{
-        const req = await productCtrl.get(productId);
+    private async getUserId(data:{productId?:number, descId?:number}):Promise<number>{
+        const {descId, productId} = data;
+        let id:number|undefined = productId;
+            
+        if(descId !== undefined){
+            const req = await this.get(id);
+            const {product_id} = (<description>req.data);
+            id = product_id;
+        }
+
+        if(!id) throw new Error('Id is undifined');
+
+        const req = await productCtrl.get(id);
         const {seller_id} = req.data;
 
         const reqSeller = await sellerCtrl.get(seller_id);
         const {user_id} = reqSeller.data;
-        return user_id;
+        return user_id;       
     }
+    
 }
 
 export default DescriptionsController;
